test(uleb128): move inline tests to integration file
